Add type None as placeholder type for no user-data
Add None type as a replacement for (). I almost would have a preference
for Empty but we've got that in use for ParameterValue, and also "empty"
is a bit too close to "empty set" and the user supplied application
state is not a set that gets built up as such. None is ok.

Change convention for `Program c a` to `Program x a`. x works as well for
"context" as c does, and all the c characters hanging around looked a
little odd.