Change Parameters to use HashMaps
Probably long overdue. Makes _using_ the parameter values a lot easier,
since the first thing you need to do before you can pattern match on the
value is lookup whether the option is there at all in the first place.

Limit the depth to which this is plumbed in the
argument parsing code; that should happen at some point.