Change output to channel of functions
Experiment with changing the output channel from a

 :: TQueue Rope

to a

 :: TQueue (Handle -> IO)

so that we can use `renderIO` when writing things that are saturated
with Ansi escape codes; prettyprinter and prettyprinter-ansi-terminal
works internally in Data.Text fragments and there's not much reason to
fight that when they supply an action which dumps the stream of their
fragments directly into an output handle.

Obviously this costs allocating a function each time someone does an
output. How does that compare to allocating the output into a Rope?