Use renderLazy directly
Reading the code for **prettyprinter-ansi-terminal**, discovered that
their `renderStrict` is just a wrapper around `renderLazy` and then
forcing to strict Data.Text. Since we have an instance of Textual to
turn a lazy Data.Text.Lazy stream into a Rope, we can just use that,
essentially only replacing the [] with a fingertree but saving a
potentially large allocation in the process.