Add Binary instance for Rope
This is a bit of a kludge. It's here, rather than Rope, because of the
circular dependency problem; Copre.Text.Bytes depends on
Core.Text.Utilities so we can't have a Textual instance. It's fine; this
is for the case where you need to flatten your text to a ByteString, and
we already have the (with caveats) Textual instance for that.