Fold from left when calculating hash of a Rope
I thought this would have to be done character-wise, but the underlying
Hashable instance for ShortText which delgates to ShortByteString seems
to correctly calculate a hash byte-wise from the left.

So all that needed to be changed was to fold the pieces of ShortText
making up the FingerTree backing a Rope from the left.

Fixes #42.