Deprecate short names
The functions `retreive` and `update` were always just "conveniences"
but they keep getting in the way and I always use `getApplicationState`
anyway.

Likewse `input` and `output` are _way_ too common and in an effort to
free up these identifiers move them to `inputEntire` and `outputEntire`
respectively.

Appropriate DEPREACATED pragmas on the old names.