Directly walk SimpleDocStream convert AnsiColour to Rope
Reimplement `render` function to directly walk the SimpleDocStream
that resulted from **prettyprinter** processing the Doc AnsiColour and
processing the AnsiColours into ANSI escape codes, appending to a Rope
as we go.

Introduce a test snippet that exercises all the colours and the
behaviour of pushing / popping tokens onto the stack.