Refactor async and catches into catches and race
The exception handling logic was severly lacking, and had what appear to
be if not race conditions then severe inconsistencies. Begin resolving
by not using catches _inside_ `async`, but instead letting **async** do
it's own thing with exceptions (which it needs to do with AsyncCancelled
for managing threads) and instead catching those it allows to escape.

Simplify `executionAction` by removing it and calling `subProgram`
directly inline.